Filtering to make sure only showing the teachers
