+++ src/main/java/com/leveluplearning/repositories/UsersRepo.java
@@ -1,8 +1,11 @@
 package com.leveluplearning.repositories;
 
 import com.leveluplearning.models.User;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 /**
  * Created by daniel on 7/6/17.
@@ -10,5 +13,8 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UsersRepo extends CrudRepository<User, Long> {
+    @Query("select u from User u where u.profSum is not null")
+    List<User> findTeachers();
+
     public User findByUsername (String username);
 }
