+++ src/main/resources/templates/views/viewAllTeacherProfiles.html
@@ -55,6 +55,8 @@
     <h1>Teachers</h1>
     <h2>Listing of all teachers available</h2>
 
+
+
     <div th:each="teacher : ${teachers}">
 
         <!--<div th:each="image : ${teacher.images}">-->
